Organize Coding Agent Files
