+++ src/os_assistant/tools/code_agent/core/models.py
