+++ src/os_assistant/tools/code_agent/wrapper.py
@@ -1,7 +1,7 @@
 from langchain_core.tools import tool
 
-from .run_code import run_code_execution
-from .parsers import ensure_string
+from .core.run_code import run_code_execution
+from .utils.parsers import ensure_string
 
 
 @tool
